refactor(Source Page): Moves the data source page into the pages folder.
