--- ui/src/containers/ExportModal/exportmodal.css
@@ -1,2 +1,0 @@
-.inputWrapper {
-  padding: 20px; }
