+++ ui/src/pages/DataSourcePage/ExportManager/ExportForm/ProjectionInput/index.js
